requirements.txt
Package            Version
------------------ ---------
bitlyshortener     0.6.0
cachetools         5.0.0
certifi            2021.10.8
charset-normalizer 2.0.12
idna               3.3
pip                20.3.4
pkg-resources      0.0.0
python-dotenv      0.19.2
requests           2.27.1
setuptools         44.1.1
urllib3            1.26.8
